Revert "chore: move Header component to components folder"
This reverts commit 311a78d88ea8ecd9f8e1fe564da417812ab66e22.
