+++ client/src/components/header/User.tsx
@@ -1,4 +1,4 @@
-import AccountDropdopwnMenu from '@/components/header/components/AccountDropdownMenu';
+import AccountDropdopwnMenu from '@/components/header/AccountDropdownMenu';
 import { useAuth } from '@/hooks';
 
 export default function User() {
