--- client/src/components/header/components/index.ts
@@ -1,6 +1,0 @@
-export { default as AccountDropdownItem } from './AccountDropdownItem';
-export { default as AccountDropdownMenu } from './AccountDropdownMenu';
-export { default as HamburgerMenu } from './HamburgerMenu';
-export { default as Navigation } from './Navigation';
-export { default as NavigationItem } from './NavigationItem';
-export { default as User } from './User';
