+++ client/src/layouts/Header.tsx
@@ -1,8 +1,10 @@
-import { Navigation, HamburgerMenu, User } from '@/components/header/components';
+import { cn } from '@/lib/utils';
+import Navigation from '@/components/header/Navigation';
+import HamburgerMenu from '@/components/header/HamburgerMenu';
 import { Button } from '@/components/ui/button';
+import User from '@/components/header/User';
 import { useScroll, useAuth } from '@/hooks/';
 import { Link } from 'react-router-dom';
-import { cn } from '@/lib/utils';
 
 function Header() {
   const { user } = useAuth();
